Fix percent-pos property
The 'percent-pos' property has to be float. Trying to retrieve this property was raising an exception, because e.g. 'int("0.0000")' raises a ValueError.

+++ mpv.py
@@ -525,7 +525,7 @@ ALL_PROPERTIES = {
         'avsync':                      (float,  'r'),
         'total-avsync-change':         (float,  'r'),
         'drop-frame-count':            (int,    'r'),
-        'percent-pos':                 (int,    'rw'),
+        'percent-pos':                 (float,  'rw'),
         'ratio-pos':                   (float,  'rw'),
         'time-pos':                    (float,  'rw'),
         'time-start':                  (float,  'r'),
